+++ .gitignore
@@ -39,3 +39,5 @@ yarn-error.*
 
 expo-env.d.ts
 # @end expo-cli
+
+.env
