+++ .vscode/settings.json
@@ -1,0 +1,16 @@
+{
+  "files.exclude": {
+    "**/.git": true,
+    "**/.svn": true,
+    "**/.hg": true,
+    "**/.DS_Store": true,
+    "**/Thumbs.db": true,
+    "**/.expo": true,
+    ".vscode": true,
+    "node_modules": true
+  },
+  "hide-files.files": [
+    ".vscode",
+    "node_modules"
+  ]
+}
