+++ index.js
@@ -5,7 +5,7 @@ import { ExpoRoot } from 'expo-router';
 
 // Must be exported or Fast Refresh won't update the context
 export function App() {
-  const ctx = require.context('./app');
+  const ctx = require.context('./src/app');
   return <ExpoRoot context={ctx} />;
 }
