+++ src/app/+not-found.tsx
@@ -1,6 +1,6 @@
 import { Link, Stack } from 'expo-router';
 import { View } from 'react-native';
-import { Text } from '~/components/ui/text';
+import { Text } from '@/components/ui/text';
 
 export default function NotFoundScreen() {
   return (
