+++ src/app/_layout.tsx
@@ -1,15 +1,15 @@
-import '~/global.css';
+import '@/global.css';
 
 import { DarkTheme, DefaultTheme, Theme, ThemeProvider } from '@react-navigation/native';
 import { Stack } from 'expo-router';
 import { StatusBar } from 'expo-status-bar';
 import * as React from 'react';
 import { Appearance, Platform, View } from 'react-native';
-import { NAV_THEME } from '~/lib/constants';
-import { useColorScheme } from '~/lib/useColorScheme';
+import { NAV_THEME } from '@/lib/constants';
+import { useColorScheme } from '@/lib/useColorScheme';
 import { PortalHost } from '@rn-primitives/portal';
-import { ThemeToggle } from '~/components/ThemeToggle';
-import { setAndroidNavigationBar } from '~/lib/android-navigation-bar';
+import { ThemeToggle } from '@/components/ThemeToggle';
+import { setAndroidNavigationBar } from '@/lib/android-navigation-bar';
 
 const LIGHT_THEME: Theme = {
   ...DefaultTheme,
