+++ src/components/ui/avatar.tsx
@@ -1,6 +1,6 @@
 import * as AvatarPrimitive from '@rn-primitives/avatar';
 import * as React from 'react';
-import { cn } from '~/lib/utils';
+import { cn } from '@/lib/utils';
 
 function Avatar({
   className,
